requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
decorator==5.1.1
defusedxml==0.7.1
Django==4.1.2
django-allauth==0.51.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
future==0.18.2
geocoder==1.38.1
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
oauthlib==3.2.1
Pillow==9.2.0
psycopg2==2.9.4
pycparser==2.21
PyJWT==2.5.0
python3-openid==3.2.0
pytz==2022.4
ratelim==0.1.6
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
types-cryptography==3.3.23
uritemplate==4.1.1
urllib3==1.26.12
